update to 26.1.3.4

+++ PKGBUILD
@@ -1,37 +1,38 @@
+# Maintainer Sienna <i at kals dev>
 # Maintainer sukanka <su975853527 at gmail dot com>
 # Maintainer: taotieren <admin@taotieren.com>
+#
+# 注意：中望CAD官方未提供固定下载链接，且 deb 二进制文件不可随意分发。
+# 请用户自行从官网下载对应版本的 amd64 deb 包，然后通过环境变量传入：
+#   ZWCAD_DEB=/path/to/zwcad2026_26.1.3.4_amd64.deb yay -S zwcad-bin
 
 pkgname='zwcad-bin'
 _pkgname='zwcad'
-pkgver=25.0.3.4
+pkgver=26.1.3.4
 _year=20${pkgver:0:2}
-pkgrel=2
+pkgrel=1
 epoch=1
 pkgdesc="ZWSOFT研发的二维CAD软件，可满足看图、审图、打印工作需要"
-# arch=('x86_64' 'aarch64')
 arch=('x86_64')
 license=('unknown')
 url="https://www.zwcad.com"
 provides=("zwcad")
+conflicts=("zwcad")
 depends=(
     bash
     curl
     expat
-    freeimage
+    fontconfig
     freetype2
     gcc-libs
-    qrencode
     glib2
     glibc
     glu
-    giflib
     hicolor-icon-theme
-    log4cpp
     libdrm
     libglvnd
     libjpeg-turbo
-    libidn11
-    libpng12
+    libpng
     libx11
     libxcb
     libxext
@@ -42,56 +43,47 @@ depends=(
     python
     krb5
     util-linux-libs
+    xdg-utils
     zlib
-    # AUR
-    openssl-1.0
-    libldap24
 )
 makedepends=('patchelf')
-source=('zwcad.sh')
-# 由于上游使用动态验证下载，每次下载中间的信息会变动，于是改为从第三方软件下载。
-# source_x86_64=("${_pkgname}-${pkgver}-x86_64.deb::https://download.zwsoft.cn/202411012038/ded6b656f41763609ece8ac461699cd3/zwcad/cad_linux/${_year}/zwcad${_year}_${pkgver}_amd64.deb")
-source_x86_64=("${_pkgname}-${pkgver}-x86_64.deb::https://aur-repo.taotieren.com:3443/atzlinux/pool/non-free/z/zwcad${_year}/zwcad${_year}_${pkgver}_amd64.deb")
-# source_aarch64=("${_pkgname}-${pkgver}-aarch64.deb::https://dl.zwsoft.cn/zwcad/cad_linux/${_year}/zwcad${_year}_${pkgver}_arm64.deb?auth_key=1690549857-0-0-5896181fcadbec59eb315d18a65cc634")
-# source_aarch64=("${_pkgname}-${pkgver}-aarch64.deb::https://aur-repo.taotieren.com:3443/atzlinux/pool/non-free/z/zwcad${_year}/zwcad${_year}_${pkgver}_arm64.deb")
-sha512sums=('138d2be8843326e7a53cd62e18c4b7173e655a61435931a4a19588da89fa9af61cc808c020e91900d3f7a60b4c5c1b720bb6442f1c0cafe226372687ed686e70')
-sha512sums_x86_64=('6545800e3cfa648fcca2d16de698ed532ab6116a30bbfbc1057aafb6d200462731b01ed5b03fa5532a3059e1c4eb47fbb07af56b69b97f535f625b663fbb19da')
+source=('zwcad.sh' 'zwlmgr.sh')
+sha512sums=('68bcd4912da5c77c1cde82645f01c01473740ae996501873141e2c2694917f1b84872ab13de4d6b0aa613e7eb2c388cc24f39ba6ebe795e72c31422bfbca91ae'
+            'ec16097d1fedb47f7bfbae91bb24823cbd62a406acbcf3c23492e17f58528053bca2ff13bcd53b33899383e35922afcbde1b237fc8546e5533fba0794668cad1')
 options=(!strip !debug)
-# no need to strip, it only decreases the installed size by 24.44 MiB.
 
 prepare() {
-    cd $srcdir
+    if [[ -z "${ZWCAD_DEB}" ]]; then
+        error "ZWCAD_DEB 环境变量未设置。"
+        msg2 "请从官网下载中望CAD Ubuntu版 deb 安装包，例如："
+        msg2 "  zwcad2026_26.1.3.4_amd64.deb"
+        msg2 "然后通过环境变量传入路径后重新构建："
+        msg2 "  ZWCAD_DEB=/path/to/zwcad${_year}_${pkgver}_amd64.deb yay -S zwcad-bin"
+        exit 1
+    fi
+    if [[ ! -f "${ZWCAD_DEB}" ]]; then
+        error "ZWCAD_DEB 指向的文件不存在：${ZWCAD_DEB}"
+        exit 1
+    fi
+
+    cd "${srcdir}"
+    ar x "${ZWCAD_DEB}" data.tar.xz
     tar -xJf data.tar.xz -C "${srcdir}"
 
-    cp zwcad.sh zwlmgr.sh
-    sed -i '$d' zwlmgr.sh
-    echo './ZwLmgr /language zh-CN /language zh-CN' >>zwlmgr.sh
-
-    cd $srcdir/opt/apps/zwcad${_year}
-    #system qt
-    rm -rf lib/libQt* plugins qt.conf
-
-    #system python
-    cd ZwPyRuntime
+    cd $srcdir/opt/apps/zwcad${_year}/ZwPyRuntime
     local _pyver=$(python -V | cut -d' ' -f2)
     test -d python${_pyver%.*} || mkdir python${_pyver%.*}
     cp python3.8/ZwPyRuntime.so python${_pyver%.*}
     rm -rf python3.{4,5,6,7,8}
 }
-build() {
 
+build() {
     cd $srcdir/opt/apps/zwcad${_year}
 
-    # Fix ./ZWLMGRRUN.sh and ./ZWCADRUN.sh not found
-    sed -i 's|ZWLMGRRUN.sh|zwlmgr\x0\x0\x0\x0\x0\x0|g' ZWCAD
-    sed -i 's|ZWCADRUN.sh|zwcad\x0\x0\x0\x0\x0\x0|g' ZwLmgr
-
-    # Fix desktop
+    # Fix desktop entries to use packaged wrappers
     cd $srcdir/usr/share/applications/
     sed -i "s|^Exec=.*|Exec=zwcad %F|g;s|^Icon=.*|Icon=zwcad${_year}|g" "zwcad${_year}.desktop"
     sed -i "s|^Exec=.*|Exec=zwlmgr|g;s|^Icon=.*|Icon=zwcad${_year}|g" "zwlmgr${_year}.desktop"
-    sed -i "s|^Exec=.*|Exec=zwcloud2d|g;s|^Icon=.*|Icon=zwcad${_year}|g" "zwcloud2d${_year}.desktop"
-
 }
 
 package() {
@@ -103,22 +95,13 @@ package() {
     # create executable
     install -Dm755 ${srcdir}/zwcad.sh "$pkgdir"/usr/bin/zwcad
     install -Dm755 ${srcdir}/zwlmgr.sh "$pkgdir"/usr/bin/zwlmgr
-    ln -s /opt/zwcad/ZwCloud2DLauncher "$pkgdir"/usr/bin/zwcloud2d
 
-    cd "$pkgdir"/opt/zwcad
-    install -Dm644 sense4_usb.rules ${pkgdir}/etc/udev/rules.d/sense4_usb.rules
+    # install udev rule
+    install -Dm644 "${pkgdir}"/opt/zwcad/sense4_usb.rules ${pkgdir}/etc/udev/rules.d/sense4_usb.rules
 
-    # fix rpath
-    find ${pkgdir}/opt/zwcad -executable -type f -print0 |
-        xargs -0 -i sh -c 'patchelf --set-rpath "\$ORIGIN:/usr/lib/zwcad" "{}" 2> /dev/null||true'
+    # remove hardcoded build rpath from binaries
+    find ${pkgdir}/opt/zwcad -maxdepth 1 -executable -type f -print0 |
+        xargs -0 -i sh -c 'patchelf --set-rpath "\$ORIGIN:\$ORIGIN/lib" "{}" 2> /dev/null || true'
 
-    # move libs
-    mkdir -p ${pkgdir}/usr/lib/zwcad
-    #     rm -rf "$pkgdir"/opt/zwcad/lib/{libcurl*,libidn*,libpng*}
-    mv "$pkgdir"/opt/zwcad/lib/lib* ${pkgdir}/usr/lib/zwcad
-    mv "$pkgdir"/opt/zwcad/lib/GL ${pkgdir}/usr/lib/zwcad/
-    # remove unused files
-    rm -rf "$pkgdir"/opt/zwcad/{Icons,ZWCADRUN.sh,ZWLMGRRUN.sh}
-    rm -rf ${pkgdir}/opt/zwcad/{libfree*,libqren*,libcrypto*,libgmp*,liblog4cpp*,libssl*,*.rules}
-    rm -rf ${pkgdir}/opt/zwcad/lib/{libicu*,libpcre*}
+    # keep bundled libs in /opt/zwcad; Qt plugins and dotnet runtime must stay intact
 }
